refactor(ui): ui component paths updated.
