+++ src/app/login/page.tsx
@@ -1,8 +1,8 @@
 'use client'
 
-import { Button } from "@/components/ui/button";
-import { Card, CardContent } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
+import { Button } from "@/ui/button";
+import { Card, CardContent } from "@/ui/card";
+import { Input } from "@/ui/input";
 import axios from "axios";
 import Link from "next/link";
 import { useEffect, useState } from "react";
